backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic[email]
argon2-cffi
passlib
python-dotenv
python-jose[cryptography]
pydantic-settings
python-multipart
ultralytics
torch 
torchvision 
segment-anything
Pillow
psycopg2-binary
python-dotenv
pytest
httpx